requirements.txt
Faker==8.8.1
kafka-python==2.0.2
pymongo==3.11.4